feat: proposal lcds
