+++ src/lib/app-provider/env.ts
@@ -82,8 +82,7 @@ export enum CELATONE_QUERY_KEYS {
   PROPOSALS_LCD = "CELATONE_QUERY_PROPOSALS_LCD",
   PROPOSAL_PARAMS = "CELATONE_QUERY_PROPOSAL_PARAMS",
   PROPOSAL_TYPES = "CELATONE_QUERY_PROPOSAL_TYPES",
-  GOV_PARAMS = "CELATONE_QUERY_GOV_PARAMS",
-  UPLOAD_ACCESS_PARAMS = "CELATONE_QUERY_UPLOAD_ACCESS_PARAMS",
+  UPLOAD_ACCESS_PARAMS_LCD = "CELATONE_QUERY_UPLOAD_ACCESS_PARAMS_LCD",
   // PUBLIC PROJECT
   PUBLIC_PROJECTS = "CELATONE_QUERY_PUBLIC_PROJECTS",
   PUBLIC_PROJECT_BY_SLUG = "CELATONE_QUERY_PUBLIC_PROJECT_BY_SLUG",
