+++ src/lib/pages/deploy/index.tsx
@@ -17,7 +17,7 @@ import { Loading } from "lib/components/Loading";
 import { Stepper } from "lib/components/stepper";
 import { UserDocsLink } from "lib/components/UserDocsLink";
 import WasmPageContainer from "lib/components/WasmPageContainer";
-import { useUploadAccessParams } from "lib/services/wasm/code";
+import { useUploadAccessParamsLcd } from "lib/services/wasm/code";
 import { AccessConfigPermission } from "lib/types";
 
 const getAlertContent = (
@@ -59,7 +59,7 @@ const Deploy = () => {
   const {
     chainConfig: { prettyName: chainPrettyName },
   } = useCelatoneApp();
-  const { data, isFetching } = useUploadAccessParams();
+  const { data, isFetching } = useUploadAccessParamsLcd();
 
   const isPermissionedNetwork =
     data?.permission !== AccessConfigPermission.EVERYBODY;
