+++ src/lib/pages/proposal-details/components/InvalidProposal.tsx
@@ -1,0 +1,5 @@
+import { InvalidState } from "lib/components/state";
+
+export const InvalidProposal = () => (
+  <InvalidState title="Proposal does not exist" />
+);
