+++ src/lib/pages/proposal-details/components/index.ts
@@ -1,3 +1,2 @@
-export * from "./proposal-overview";
-export * from "./proposal-top";
-export * from "./vote-details";
+export * from "./proposal-details-body";
+export * from "./InvalidProposal";
